// apps/api/src/common/constants/error-codes.ts
export enum ErrorCode {
  VALIDATION_ERROR = 'VALIDATION_ERROR',
  UNAUTHORIZED = 'UNAUTHORIZED',
  FORBIDDEN = 'FORBIDDEN',
  INVALID_CREDENTIALS = 'INVALID_CREDENTIALS',
  INVALID_REFRESH_TOKEN = 'INVALID_REFRESH_TOKEN',
  EMAIL_VERIFICATION_TOKEN_INVALID = 'EMAIL_VERIFICATION_TOKEN_INVALID',
  EMAIL_ALREADY_VERIFIED = 'EMAIL_ALREADY_VERIFIED',
  EMAIL_NOT_VERIFIED = 'EMAIL_NOT_VERIFIED',
  PASSWORD_RESET_TOKEN_INVALID = 'PASSWORD_RESET_TOKEN_INVALID',
  CANNOT_BLOCK_SELF = 'CANNOT_BLOCK_SELF',
  INVALID_IMAGE_FILE = 'INVALID_IMAGE_FILE',
  FILE_TOO_LARGE = 'FILE_TOO_LARGE',
  USER_NOT_FOUND = 'USER_NOT_FOUND',
  POST_NOT_FOUND = 'POST_NOT_FOUND',
  COMMENT_NOT_FOUND = 'COMMENT_NOT_FOUND',
  BOARD_NOT_FOUND = 'BOARD_NOT_FOUND',
  CATEGORY_NOT_FOUND = 'CATEGORY_NOT_FOUND',
  CATEGORY_SLUG_ALREADY_EXISTS = 'CATEGORY_SLUG_ALREADY_EXISTS',
  BOARD_SLUG_ALREADY_EXISTS = 'BOARD_SLUG_ALREADY_EXISTS',
  CATEGORY_HAS_BOARDS = 'CATEGORY_HAS_BOARDS',
  BOARD_HAS_POSTS = 'BOARD_HAS_POSTS',
  ATTACHMENT_NOT_FOUND = 'ATTACHMENT_NOT_FOUND',
  ATTACHMENT_ALREADY_LINKED = 'ATTACHMENT_ALREADY_LINKED',
  TOO_MANY_TAGS = 'TOO_MANY_TAGS',
  POST_NOT_DELETED = 'POST_NOT_DELETED',
  COMMENT_NOT_DELETED = 'COMMENT_NOT_DELETED',
  INVALID_PARENT_COMMENT = 'INVALID_PARENT_COMMENT',
  PARENT_POST_DELETED = 'PARENT_POST_DELETED',
  INVALID_REACTION_TARGET = 'INVALID_REACTION_TARGET',
  REACTION_TYPE_NOT_SUPPORTED = 'REACTION_TYPE_NOT_SUPPORTED',
  NOTIFICATION_NOT_FOUND = 'NOTIFICATION_NOT_FOUND',
  REPORT_NOT_FOUND = 'REPORT_NOT_FOUND',
  ALREADY_REPORTED = 'ALREADY_REPORTED',
  CANNOT_REPORT_SELF = 'CANNOT_REPORT_SELF',
  INVALID_REPORT_TARGET = 'INVALID_REPORT_TARGET',
  REPORT_ALREADY_RESOLVED = 'REPORT_ALREADY_RESOLVED',
  ROLE_NOT_FOUND = 'ROLE_NOT_FOUND',
  CANNOT_MODIFY_SELF_ROLE = 'CANNOT_MODIFY_SELF_ROLE',
  CANNOT_GRANT_SUPER_ADMIN = 'CANNOT_GRANT_SUPER_ADMIN',
  BANNED_WORD_ALREADY_EXISTS = 'BANNED_WORD_ALREADY_EXISTS',
  BANNED_WORD_NOT_FOUND = 'BANNED_WORD_NOT_FOUND',
  IP_ALREADY_BANNED = 'IP_ALREADY_BANNED',
  IP_BAN_NOT_FOUND = 'IP_BAN_NOT_FOUND',
  SETTING_NOT_FOUND = 'SETTING_NOT_FOUND',
  AD_SLOT_NOT_FOUND = 'AD_SLOT_NOT_FOUND',
  AD_NOT_FOUND = 'AD_NOT_FOUND',
  CONTAINS_BANNED_WORD = 'CONTAINS_BANNED_WORD',
  RESOURCE_NOT_FOUND = 'RESOURCE_NOT_FOUND',
  EMAIL_ALREADY_EXISTS = 'EMAIL_ALREADY_EXISTS',
  NICKNAME_ALREADY_EXISTS = 'NICKNAME_ALREADY_EXISTS',
  ALREADY_REACTED = 'ALREADY_REACTED',
  RATE_LIMIT_EXCEEDED = 'RATE_LIMIT_EXCEEDED',
  INTERNAL_SERVER_ERROR = 'INTERNAL_SERVER_ERROR',
}